"keepAlive" in MQTT is 2 byte (quint16)
http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc385349238

+++ src/qmqtt_client.cpp
@@ -118,13 +118,13 @@ void QMQTT::Client::setPassword(const QString& password)
     d->setPassword(password);
 }
 
-int QMQTT::Client::keepAlive() const
+quint16 QMQTT::Client::keepAlive() const
 {
     Q_D(const Client);
     return d->keepAlive();
 }
 
-void QMQTT::Client::setKeepAlive(const int keepAlive)
+void QMQTT::Client::setKeepAlive(const quint16 keepAlive)
 {
     Q_D(Client);
     d->setKeepAlive(keepAlive);
